db: remove some historic views and update scripts over a year old
Older explicit upgrade scripts might not work as they jump from one
version to the next.

The most recent one of these only updates the download count for a
transfer from historic values in the auditlogs. If you are jumping
over that to the current version then the download counts might not be
accurate right away unless you migrate via the 2.49 version.