D3UI3 disable user theme until it is updated. (#2272)
Some issues with themes have been reported. Some sites disable them by
defualt. Perhaps we should make that the default until the themese
have been updated a little bit.

It is best not to present a UI being "unable to log in and unable to
upload files". Also a catch22, if you can not login then you can not
change your theme.