Remove old script which might be confusing for some users. (#2296)
The upgrade.php has not changed in around 10 years and has a versions
that references a 1.5 release. Without re-validating that the
upgrade.php is working as expected on the current 2.x series release
it is a bit of a liability being here. I have also added a README.md
in there to direct folks as to what is interesting in this directory
and mention the "view drop" permission issue for MariaDB and direct
folks to the right part of the documentation for that issue.