D3UI3: update exception and transfer expired page style
Make both of these pages centered, use the bootstrap danger box, add
an icon to the mix to indicate an issue. This should make both of
these cases very obvious that an error has occurred.

You can see an exception page by visiting the download page and
mangling the token in the URL bar, for example adding something to the
end of it.

You can see the transfer not found page by visiting an existing
transfer details and mangling the transfer_id to be some much larger
number than the one you are viewing.