UX: On OSX/Safari it seems lengthComputable is not set. (#2324)
The evt.lengthComputable is false on osx/safari. This leads to the
progressbar not being updated in the UI. The code doesn't actually
need to use the total size from the evt message so if there is some
indication of what is done we can update the progress anyway.

This should allow a nice download progress message on OSX/Safari.

Tested in an A/B on the current version of macos/safari as of today.