d3ui3 doing a notify and reload on language selection (#2325)
It would be cleaner here to be able to do a reload right away and then
perform the notify in that reloaded window. That would need some more
verification for example that you came from the same page and only
allowing some messages to be sent. It is a bit of a security quagmire
so just notifying that things went well and then redirecting soon
after seems like the simpler short term solution.

This may be also what is effecting the language selection that some
people have reported since the notify/reload issue was not used in the
lang selection in 2.x.