add basic ensureLocalIdPMetadata for Shibboleth (#2335)
For this to be fully functional in the new stats update I need to play
around more with some IdPs for Shibboleth and check that per IdP data
is properly presented. Many of the fields in the idps table for
Shibboleth will be left as null for now.

For SSP the code was reading the remote idp listing from the library.
For the code to match in Shibboleth it should also be looking at the
metadata for the remote IdPs that the Shibboleth SP knows about and
adding that information into the local idps table in FileSender.