Auth shibboleth: add IdP import codepath for shibboleth (#2336)
This is an update to my initial update for
https://github.com/filesender/filesender/issues/2334

With this the admin has a potential option to populate the IdPs table
with metadata about the IdPs that are available in their Shibboleth SP
setup.

This is only tested on a small local setup. More work and updates are
likely to be needed for a live setup with real data.