Config: SAML additional attriubtes to config (#2350) (#2351)
* Config: SAML additional attriubtes to config (#2350)

* Config: SAML additional attriubtes to config

A new auth_config_additional_regex_files config key that matches
against the whole array (one at a time) for the picked attributes.

Perhaps we should also add at some stage a "sort" type key to allow
the provided array to be sorted so that the order that multiple
matching config files is applied is determined explicitly. For
example, student and employee we might like the student to sort first
so that the employee config can be more permissive or have a larger
quota.

* catch bad config cases and tell admin in strong terms

* switch order

* update v3 docs

---------

Co-authored-by: monkeyiq <monkeyiq@users.sourceforge.net>