ui: refactor download page to use new common code block
Note that the email verification window still uses jquery-ui for the
dialog. Trying to move from that to bootstrap doesn't seem to work as
bootstrap is not fully in use any more. I guess that will have to be
moved to more custom non framework code in the future to have a
uniform close button etc.

The email verification feature should now work again.

config.php:

$config['download_verification_code_enabled'] = true;