ui3 option to allow guest to see their uploads
This is to address
https://github.com/filesender/filesender/issues/2046

note that we do not cater for guests who have already expired here.
This is a good first step to allow a guest to check if their upload is
in the system which seems like a very useful feature for a guest who
is submitting files and wishes to double check that they were sent
correctly (from FileSender's perspective).