db: add an index on files.uid
This addresses https://github.com/filesender/filesender/issues/2407

For postgresql leaving the uid column as a string the index for 0.5m
files is about 58mb. Using the native uuid type for a column that is a
copy of the uid column the same index is 17mb. So a migration to the
native uuid types in mariadb and postgresql is probably a good idea at
some stage.