db: precache files.download_count
This offers significant performance improvements for a my transfers
page with transfers containing 10,000 files each. If you have a number
of such transfers the load time on that page might got from 2 minutes
to less than 10 seconds.

The migration might take a moment as it will run this somewhat
expensive query on the auditlog table to find out how many downloads
have happened for each file in the database. It is much, much more
efficient to precache that count so it is worth doing this the one
time for the pleasure of faster load times from then on out.

This relates to https://github.com/filesender/filesender/issues/2405
but does not include the transfer_id (trid) value in the auditlog as
yet. That will be a subsequent investigation.

Note that you will have to run the database.php script to create the
files.download_count column and it should populate that column at
database.php execution time which will take some time.