perf: lookup File object by id if possible
If we are making a custom URL format we might as well tag the file->id
as a fragment id to pass to the stream. The stream can then lookup the
file using the numeric id primary key which will be much faster than
on a uuidv4 value.

This is mainly server side load when making archives for non encrypted
transfers. It will probably make minimal performance impact for other
things.