Split public and storage uuid in File class
The puid is the "public" uuid which is version 4.
The old uid is now using version 7 (temporal) uuid generation.

The file uid is not sent to the browser. The new puid is used instead.
Files are stored using the uid which now becomes a temporal uuid. That
allows for the filesystem or database doing the storage of data to
enjoy a performance boost in indexing and lookup. While the uuid that
might be sent and used in the browser is a uuidv4 in the form of the
puid.

This is a follow up to work done a while ago where uuidv7 was
considered. In this updated PR the file.uid is designed to be a server
only identifier. The puid is the public uuid which remains a uuidv4
identifier to retain the full entropy.