db perf: allow bulk loading of new files when creating transfer
This adds an option to use the database bulk load API to create new
File records. This can be time consuming for transfers with 10k
files or more. This option is off by default and can be turned on my
setting the threshold which you would like to switch to the bulk api.

Note that doing this means that UUID generation is not checked for
duplicates.

To enable this you have to set in config.php something like

'create_transfer_uses_bulk_insert_threshold' => 500,